Move Kafka router into a separate module
When the Kafka router was moved into the factory so that it could
be used by background tasks, this cased the Kafka handler module
to never be imported, so no handler functions were ever registered.
Use the same pattern as Ook and move the Kafka router into its own
module to avoid circular dependencies.
