+++ src/qservkafka/handlers/kafka.py
@@ -1,4 +1,10 @@
-"""Kafka router and consumers."""
+"""Kafka router and consumers.
+
+The `kafka_router` symbol must be imported from this module, not from its true
+source at `~qservkafka.kafkarouters.kafka_router`, to ensure that the router
+is properly configured with its subscribers and publishers. Otherwise, the
+router will have no subscribers and no publishers and will thus do nothing.
+"""
 
 from typing import Annotated
 
@@ -6,7 +12,7 @@ from fastapi import Depends
 
 from ..config import config
 from ..dependencies.context import ConsumerContext, context_dependency
-from ..factory import kafka_router
+from ..kafkarouters import kafka_router
 from ..models.kafka import JobRun, JobStatus
 
 publisher = kafka_router.publisher(config.job_status_topic)
